% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load data.R
\name{get_and_combine}
\alias{get_and_combine}
\title{Get and combine data for MT and BS (gas fun)}
\usage{
get_and_combine()
}
\value{
Returns data frame
}
\description{
"Labor must work harder to attract and retain members." - BS
"Anyone who thinks it's smart to cut immigration is sentencing Australia to poverty" - MT
}
