% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load data.R
\name{sentences_BS}
\alias{sentences_BS}
\title{Get data for BS speeches as sentences}
\usage{
sentences_BS(path = "/TrendLock/data/speeches/sentences_BS.csv")
}
\arguments{
\item{path}{Defaults to storage}
}
\value{
Returns data frame
}
\description{
"Labor must work harder to attract and retain members." - BS
}
