% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load data.R
\name{sentences_MT}
\alias{sentences_MT}
\title{Get data for MT speeches as sentences}
\usage{
sentences_MT(path = "/Dropbox/TrendLock/data/speeches/sentences_MT.csv")
}
\arguments{
\item{path}{Defaults to storage}
}
\value{
Returns data frame
}
\description{
"Anyone who thinks it's smart to cut immigration is sentencing Australia to poverty" - MT
}
