% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load data.R
\name{word_BS}
\alias{word_BS}
\title{Get data for BS speeches as words}
\usage{
word_BS(path = "/Dropbox/TrendLock/data/speeches/word_BS.csv")
}
\arguments{
\item{path}{Defaults to storage}
}
\value{
Returns data frame
}
\description{
"Labor must work harder to attract and retain members." - BS
}
