% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load data.R
\name{word_comb}
\alias{word_comb}
\title{Get combine data for BS and MT}
\usage{
word_comb(path = "/Dropbox/TrendLock/data/speeches/words_comb.csv")
}
\arguments{
\item{path}{Defaults to storage}
}
\value{
Returns data frame
}
\description{
Blah
}
