% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load data.R
\name{word_MT}
\alias{word_MT}
\title{Get data for MT speeches as words}
\usage{
word_MT(path = "/Dropbox/TrendLock/data/speeches/word_MT.csv")
}
\arguments{
\item{path}{Defaults to storage}
}
\value{
Returns data frame
}
\description{
"Anyone who thinks it's smart to cut immigration is sentencing Australia to poverty" - MT
}
